# docs/swagger.yaml
definitions:
  github_com_Orken1119_HelpNet_internal_models.Certificate:
    properties:
      certificate_url:
        type: string
      id:
        type: integer
    type: object
...
  github_com_Orken1119_HelpNet_internal_models.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
      metadata:
        $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.Properties'
    type: object
  github_com_Orken1119_HelpNet_internal_models.ErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorDetail'
        type: array
    type: object
  github_com_Orken1119_HelpNet_internal_models.Event:
    properties:
      address:
        type: string
      end_data:
        type: string
      event_name:
        type: string
      finished:
        type: boolean
      id:
        type: integer
      information:
        type: string
      members_count:
        type: integer
      members_info:
        items:
          $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.VolunteerMainInfo'
        type: array
      neccessary_people_count:
        type: integer
      organization_id:
        type: integer
      poster_url:
        type: string
      preview_url:
        type: string
      skill_direction:
        type: string
      start_data:
        type: string
    type: object
...
  github_com_Orken1119_HelpNet_internal_models.OrganizationProfile:
    properties:
      city:
        type: string
      direction:
        type: string
      email:
        type: string
      events:
        items:
          $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.Event'
        type: array
      finished_ivents:
        items:
          $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.Event'
        type: array
      id:
        type: integer
      information:
        type: string
      name:
        type: string
      phoneNumber:
        type: string
      poster_url:
        type: string
      years_of_experience:
        type: integer
    type: object
...
  github_com_Orken1119_HelpNet_internal_models.Password:
    properties:
      confirmPassword:
        type: string
      password:
        type: string
    type: object
  github_com_Orken1119_HelpNet_internal_models.Properties:
    properties:
      additionalProp1:
        type: string
      additionalProp2:
        type: string
      additionalProp3:
        type: string
    type: object
...
  github_com_Orken1119_HelpNet_internal_models.SuccessResponse:
    properties:
      metadata:
        $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.Properties'
      result: {}
    type: object
  github_com_Orken1119_HelpNet_internal_models.VolunteerMainInfo:
    properties:
      age:
        type: integer
      city:
        type: string
      direction:
        type: string
      email:
        type: string
      grade:
        type: integer
      id:
        type: integer
      name:
        type: string
      phoneNumber:
        type: string
      photo_url:
        type: string
      skills:
        type: string
    type: object
  github_com_Orken1119_HelpNet_internal_models.VolunteerProfile:
    properties:
      age:
        type: integer
      certificates:
        items:
          $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.Certificate'
        type: array
      city:
        type: string
      direction:
        type: string
      email:
        type: string
      events_now:
        items:
          $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.Event'
        type: array
      finished:
        items:
          $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.Event'
        type: array
      grade:
        type: integer
      id:
        type: integer
      name:
        type: string
      phoneNumber:
        type: string
      photo_url:
        type: string
      skills:
        type: string
    type: object
  github_com_Orken1119_HelpNet_internal_models.VolunteerProfileEditing:
    properties:
      age:
        type: integer
      city:
        type: string
      direction:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phoneNumber:
        type: string
      photo_url:
        type: string
      skills:
        type: string
    type: object
...
host: localhost:2000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: HelpNet API
  version: "1.0"
paths:
...
  /organizations/profile/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.SuccessResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.OrganizationProfile'
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - organization
  /user/add-certificate:
    post:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.Certificate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.SuccessResponse'
            - properties:
                result:
                  type: string
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - user
  /user/change-password:
    put:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.Password'
      produces:
      - application/json
      responses:
        "200":
          description: Password successfully changed
          schema:
            additionalProperties:
              type: string
            type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - user
  /user/change-password-for-org:
    put:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.Password'
      produces:
      - application/json
      responses:
        "200":
          description: Password successfully changed
          schema:
            additionalProperties:
              type: string
            type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - user
  /user/delete-certificate/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: certificate was deleted
          schema:
            additionalProperties:
              type: string
            type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - user
  /user/edit-profile:
    put:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.VolunteerProfileEditing'
      produces:
      - application/json
      responses:
        "200":
          description: Personal data was changed
          schema:
            additionalProperties:
              type: string
            type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - user
  /user/get-all-volunteers-profile:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.SuccessResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.VolunteerProfile'
                  type: array
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - user
  /user/profile/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.SuccessResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.VolunteerProfile'
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/github_com_Orken1119_HelpNet_internal_models.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - user
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"